backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
pytest==8.3.3
pytest-django==4.9.0